feat(kodiak): testing plain text (#9)
Description
Add album sorting by album index, name and created date.
Remove query params when navigating between pages, as requested by PM. This is the same behavior as samples.
For now, the back button only redirects to /library, but it should redirect to the previous page, hopefully keeping the original search or filtering as well as correctly navigating to shared with me or my files. We have a bug opened for this : PROJ-3246 and we should handle the correct redirection there.
Screenshots/Videos
Checklist

 Contains no duplicate/temporary code
 Contains no sensitive information
 Error handling added
 Unit tests added

JIRA Issue
PROJ-2128